+++ examples/linewidth.py
@@ -1,0 +1,12 @@
+from easygraphics import *
+
+init_graph(800, 600)
+
+line(50, 50, 400, 400)
+
+set_line_width(20)
+
+line(50, 400, 400, 50)
+
+pause()
+close_graph()